renamed classes for clarity
